rename movie item
