+++ src/movies/movie-item/movie-item.scss
